administrative api and tools
Contains miscellaneous (maintenance) features:

 - support for cache slice fusion, index regeneration and object
   size caching;
 - non-commit object generation refactored to take advantage of
   'size' field
 - porcelain updated to support feature additions

The beginnings of integration into git are present in this patch,
mainly centered on caching object size; the object generation is
refactored to more elegantly exploit this.  Fusion allows smaller
(incremental) slices to be coagulated into a larger slice, reducing
overhead, while index regeneration enables repair or cleaning of the
cache index.

Note that tests for these features are included in the following
patch, as they take advantage of the rev-cache's integration into the
revision walker.

Signed-off-by: Nick Edelen <sirnot@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>