object name support
An update to caching mechanism, allowing path names to be cached for
blob and tree objects.  A list of names appearing in each cache slice
is appended to the end of the slice, which is referenced by
variable-sized indexes per entry.

This allows pack-objects to more intelligently schedule
unpacked/poorly packed object, and enables proper duplication of
rev-list's behaivor.

Signed-off-by: Nick Edelen <sirnot@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>